chore: refactor, move utils pda stuff to its own module
