+++ programs/axelar-solana-gas-service/src/processor/native.rs
@@ -1,7 +1,10 @@
 use crate::assert_valid_config_pda;
 use crate::state::Config;
 use axelar_solana_gas_service_events::event_prefixes;
-use program_utils::{transfer_lamports, validate_system_account_key, BytemuckedPda, ValidPDA};
+use program_utils::{
+    pda::{BytemuckedPda, ValidPDA},
+    transfer_lamports, validate_system_account_key,
+};
 use solana_program::account_info::{next_account_info, AccountInfo};
 use solana_program::entrypoint::ProgramResult;
 use solana_program::log::sol_log_data;
