+++ programs/axelar-solana-gateway/src/lib.rs
@@ -7,7 +7,7 @@ pub mod state;
 
 pub use bytemuck;
 pub use num_traits;
-pub use program_utils::BytemuckedPda;
+pub use program_utils::pda::BytemuckedPda;
 
 // Export current sdk types for downstream users building with a different sdk
 // version.
