+++ programs/axelar-solana-gateway/src/processor/commit_message_payload.rs
@@ -1,7 +1,7 @@
 use super::Processor;
 use crate::state::incoming_message::IncomingMessage;
 use crate::state::message_payload::MutMessagePayload;
-use program_utils::{BytemuckedPda, ValidPDA};
+use program_utils::pda::{BytemuckedPda, ValidPDA};
 use solana_program::account_info::{next_account_info, AccountInfo};
 use solana_program::entrypoint::ProgramResult;
 use solana_program::program_error::ProgramError;
