+++ programs/axelar-solana-gateway/src/state/signature_verification.rs
@@ -9,7 +9,7 @@ use bitvec::order::Lsb0;
 use bitvec::slice::BitSlice;
 use bitvec::view::BitView;
 use bytemuck::{Pod, Zeroable};
-use program_utils::BytemuckedPda;
+use program_utils::pda::BytemuckedPda;
 
 use crate::error::GatewayError;
