+++ programs/axelar-solana-gateway/tests/integration/transfer_operatorship.rs
@@ -8,7 +8,7 @@ use axelar_solana_gateway_test_fixtures::{
     SolanaAxelarIntegration, SolanaAxelarIntegrationMetadata,
 };
 use num_traits::ToPrimitive as _;
-use program_utils::BytemuckedPda;
+use program_utils::pda::BytemuckedPda;
 use solana_program_test::tokio::fs;
 use solana_program_test::{tokio, ProgramTest};
 use solana_sdk::account::ReadableAccount;
