+++ programs/axelar-solana-its/src/executable.rs
@@ -4,7 +4,7 @@ use axelar_executable::AxelarMessagePayload;
 use axelar_solana_gateway::state::message_payload::ImmutMessagePayload;
 use borsh::{BorshDeserialize, BorshSerialize};
 use interchain_token_transfer_gmp::GMPPayload;
-use program_utils::BorshPda;
+use program_utils::pda::BorshPda;
 use solana_program::account_info::{next_account_info, AccountInfo};
 use solana_program::msg;
 use solana_program::program_error::ProgramError;
