+++ programs/axelar-solana-its/src/processor/mod.rs
@@ -3,7 +3,10 @@
 use borsh::BorshDeserialize;
 use event_utils::Event as _;
 use interchain_token::process_mint;
-use program_utils::{validate_system_account_key, BorshPda, ValidPDA};
+use program_utils::{
+    pda::{BorshPda, ValidPDA},
+    validate_system_account_key,
+};
 use role_management::instructions::RoleManagementInstructionInputs;
 use role_management::processor::{
     ensure_signer_roles, ensure_upgrade_authority, RoleManagementAccounts,
