+++ programs/axelar-solana-its/src/state/mod.rs
@@ -4,7 +4,7 @@
 use std::collections::HashSet;
 
 use borsh::{BorshDeserialize, BorshSerialize};
-use program_utils::BorshPda;
+use program_utils::pda::BorshPda;
 
 pub mod deploy_approval;
 pub mod flow_limit;
